requirements.txt
streamlit
qrcode[pil]
opencv-python-headless
pandas
bcrypt
pillow